requirements.txt
fastapi
uvicorn
mysql
PyMySQL
aiomysql
pydantic
sqlalchemy
python-dotenv
cryptography
pandas
openpyxl
python-multipart
weasyprint
pypdf
jinja2
aiofiles
schedule
celery
# celery[redis]
redis
